Main: subdivided plane rendering

+++ src/model/Mesh.type.ts
@@ -4,5 +4,5 @@ export default interface Mesh {
     vertices : Float32Array;
     normals : Float32Array,
     uvs : Float32Array,
-    indices : Float32Array
+    indices : Uint16Array
 }
